+++ src/SFA.DAS.Payments.Monitoring.Jobs.JobsProxyService/JobsProxyService.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Fabric;
 using Autofac;
@@ -7,34 +7,27 @@ using Microsoft.ServiceFabric.Services.Runtime;
 using SFA.DAS.Payments.Application.Infrastructure.Logging;
 using SFA.DAS.Payments.ServiceFabric.Core;
 
-namespace SFA.DAS.Payments.Monitoring.JobsService
+namespace SFA.DAS.Payments.Monitoring.Jobs.JobsProxyService
 {
-    /// <summary>
-    /// An instance of this class is created for each service instance by the Service Fabric runtime.
-    /// </summary>
-    public class JobsService : StatelessService
+    public class JobsProxyService : StatelessService
     {
         private readonly ILifetimeScope lifetimeScope;
         private readonly IPaymentLogger logger;
 
-        public JobsService(StatelessServiceContext context, ILifetimeScope lifetimeScope, IPaymentLogger logger)
+        public JobsProxyService(StatelessServiceContext context, ILifetimeScope lifetimeScope, IPaymentLogger logger)
             : base(context)
         {
             this.lifetimeScope = lifetimeScope ?? throw new ArgumentNullException(nameof(lifetimeScope));
             this.logger = logger ?? throw new ArgumentNullException(nameof(logger));
         }
 
-        /// <summary>
-        /// Optional override to create listeners (e.g., TCP, HTTP) for this service replica to handle client or user requests.
-        /// </summary>
-        /// <returns>A collection of listeners.</returns>
         protected override IEnumerable<ServiceInstanceListener> CreateServiceInstanceListeners()
         {
-            logger.LogInfo("Creating Service Instance Listeners For Job Service");
+            logger.LogInfo("Creating Service Instance Listeners For Job Proxy Service");
 
             return new List<ServiceInstanceListener>
             {
-                new ServiceInstanceListener(context =>lifetimeScope.Resolve<IStatelessEndpointCommunicationListener>())
+                new ServiceInstanceListener(context => lifetimeScope.Resolve<IStatelessEndpointCommunicationListener>())
             };
         }
     }
