+++ src/SFA.DAS.Payments.Monitoring.Jobs.JobsProxyService/PackageRoot/ServiceManifest.xml
@@ -1,5 +1,5 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ServiceManifest Name="SFA.DAS.Payments.Monitoring.JobsServicePkg"
+<?xml version="1.0" encoding="utf-8"?>
+<ServiceManifest Name="SFA.DAS.Payments.Monitoring.Jobs.JobsProxyServicePkg"
                  Version="1.0.0"
                  xmlns="http://schemas.microsoft.com/2011/01/fabric"
                  xmlns:xsd="http://www.w3.org/2001/XMLSchema"
@@ -7,19 +7,19 @@
   <ServiceTypes>
     <!-- This is the name of your ServiceType. 
          This name must match the string used in the RegisterServiceAsync call in Program.cs. -->
-    <StatelessServiceType ServiceTypeName="SFA.DAS.Payments.Monitoring.JobsServiceType" />
+    <StatelessServiceType ServiceTypeName="SFA.DAS.Payments.Monitoring.Jobs.JobsProxyServiceType" />
   </ServiceTypes>
 
   <!-- Code package is your service executable. -->
   <CodePackage Name="Code" Version="1.0.0">
     <EntryPoint>
       <ExeHost>
-        <Program>SFA.DAS.Payments.Monitoring.JobsService.exe</Program>
+        <Program>SFA.DAS.Payments.Monitoring.Jobs.JobsProxyService.exe</Program>
       </ExeHost>
     </EntryPoint>
   </CodePackage>
 
-  <!-- Config package is the contents of the Config directory under PackageRoot that contains an 
+  <!-- Config package is the contents of the Config directoy under PackageRoot that contains an 
        independently-updateable and versioned set of custom configuration settings for your service. -->
   <ConfigPackage Name="Config" Version="1.0.0" />
