+++ src/SFA.DAS.Payments.Monitoring.Jobs.JobsService.Interfaces/IJobsService.cs
@@ -1,14 +1,14 @@
-﻿using System;
 using System.Threading.Tasks;
 using Microsoft.ServiceFabric.Actors;
 using Microsoft.ServiceFabric.Actors.Remoting.FabricTransport;
 using Microsoft.ServiceFabric.Services.Remoting;
+using SFA.DAS.Payments.Monitoring.Jobs.Messages.Commands;
 
 [assembly: FabricTransportActorRemotingProvider(RemotingListenerVersion = RemotingListenerVersion.V2_1, RemotingClientVersion = RemotingClientVersion.V2_1)]
-namespace SFA.DAS.Payments.Monitoring.JobsStatusService.Interfaces
+namespace SFA.DAS.Payments.Monitoring.Jobs.JobsService.Interfaces
 {
-    public interface IJobsStatusService : IActor
+    public interface IJobsService : IActor
     {
-        Task<(bool Finished, DateTimeOffset? endTime)> JobStepsCompleted(long jobId);
+        Task RecordEarningsJob(RecordEarningsJob message);
     }
 }
