+++ src/SFA.DAS.Payments.Monitoring.Jobs.JobsService/PackageRoot/Config/Settings.xml
@@ -1,12 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Settings xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns="http://schemas.microsoft.com/2011/01/fabric">
-  <Section Name="JobsStatusServiceActorServiceReplicatorConfig">
-    <Parameter Name="ReplicatorEndpoint" Value="JobsStatusServiceActorServiceReplicatorEndpoint" />
-    <Parameter Name="BatchAcknowledgementInterval" Value="0.005" />
-  </Section>
-  <Section Name="JobsStatusServiceActorServiceReplicatorSecurityConfig">
-    <Parameter Name="CredentialType" Value="None" />
-  </Section>
   <!-- The content will be generated during build -->
   <Section Name="ConnectionStringsSection">
     <Parameter Name="StorageConnectionString" Value="UseDevelopmentStorage=true" />
@@ -16,12 +9,18 @@
     <Parameter Name="PaymentsMonitoringConnectionString" Value="" />
   </Section>
   <Section Name="Settings">
-    <Parameter Name="EndpointName" Value="sfa-das-payments-monitoring-jobsstatus" />
-    <Parameter Name="FailedMessagesQueue" Value="sfa-das-payments-monitoring-jobsstatus-errors" />
+    <Parameter Name="EndpointName" Value="sfa-das-payments-monitoring-jobs" />
+    <Parameter Name="FailedMessagesQueue" Value="sfa-das-payments-monitoring-jobs-errors" />
     <Parameter Name="ApplicationInsightsInstrumentationKey" Value="" />
-    <Parameter Name="FailureTimeoutInSeconds" Value="150" />
-    <Parameter Name="IntervalInSeconds" Value="10" />
+    <Parameter Name="DelayedRetryTimeInSeconds" Value="120" />
     <Parameter Name="LogLevel" Value="" />
     <Parameter Name="ConsoleLogLevel" Value="" />
+  </Section>
+  <Section Name="JobsServiceActorServiceReplicatorConfig">
+    <Parameter Name="ReplicatorEndpoint" Value="JobsServiceActorServiceReplicatorEndpoint" />
+    <Parameter Name="BatchAcknowledgementInterval" Value="0.005" />
+  </Section>
+  <Section Name="JobsServiceActorServiceReplicatorSecurityConfig">
+    <Parameter Name="CredentialType" Value="None" />
   </Section>
 </Settings>
