+++ src/SFA.DAS.Payments.Monitoring.Jobs.JobsService/PackageRoot/ServiceManifest.xml
@@ -1,0 +1,34 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<ServiceManifest xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema" Name="SFA.DAS.Payments.Monitoring.Jobs.JobsServicePkg" Version="1.0.0" xmlns="http://schemas.microsoft.com/2011/01/fabric">
+  <ServiceTypes>
+    <StatefulServiceType ServiceTypeName="JobsServiceActorServiceType">
+      <Extensions>
+        <Extension Name="__GeneratedServiceType__" GeneratedId="cb971f61-6bb9-4f53-a2d2-c61656f04754|Volatile">
+          <GeneratedNames xmlns="http://schemas.microsoft.com/2015/03/fabact-no-schema">
+            <DefaultService Name="JobsServiceActorService" />
+            <ReplicatorEndpoint Name="JobsServiceActorServiceReplicatorEndpoint" />
+            <ReplicatorConfigSection Name="JobsServiceActorServiceReplicatorConfig" />
+            <ReplicatorSecurityConfigSection Name="JobsServiceActorServiceReplicatorSecurityConfig" />
+            <StoreConfigSection Name="JobsServiceActorServiceLocalStoreConfig" />
+            <ServiceEndpointV2_1 Name="JobsServiceActorServiceEndpointV2_1" />
+          </GeneratedNames>
+        </Extension>
+      </Extensions>
+    </StatefulServiceType>
+  </ServiceTypes>
+  <CodePackage Name="Code" Version="1.0.0">
+    <EntryPoint>
+      <ExeHost>
+        <Program>SFA.DAS.Payments.Monitoring.Jobs.JobsService.exe</Program>
+      </ExeHost>
+    </EntryPoint>
+  </CodePackage>
+  <ConfigPackage Name="Config" Version="1.0.0" />
+  <Resources>
+    <Endpoints>
+      <Endpoint Name="JobsServiceActorServiceEndpointV2_1" />
+      <Endpoint Name="JobsServiceActorServiceReplicatorEndpoint" />
+    </Endpoints>
+  </Resources>
+  <!-- The content will be generated during build -->
+</ServiceManifest>
