--- src/SFA.DAS.Payments.Monitoring.JobsService/JobsService.cs
@@ -1,35 +1,0 @@
-using System;
-using System.Threading.Tasks;
-using Microsoft.ServiceFabric.Actors;
-using Microsoft.ServiceFabric.Actors.Runtime;
-using SFA.DAS.Payments.Monitoring.JobsService.Interfaces;
-using SFA.DAS.Payments.Monitoring.Jobs.Messages.Commands;
-
-namespace SFA.DAS.Payments.Monitoring.JobsService
-{
-    [StatePersistence(StatePersistence.Volatile)]
-    public class JobsService : Actor, IJobsService
-    {
-        public JobsService(ActorService actorService, ActorId actorId) 
-            : base(actorService, actorId)
-        {
-        }
-
-        protected override Task OnActivateAsync()
-        {
-            ActorEventSource.Current.ActorMessage(this, "Actor activated.");
-
-            // The StateManager is this actor's private state store.
-            // Data stored in the StateManager will be replicated for high-availability for actors that use volatile or persisted state storage.
-            // Any serializable object can be saved in the StateManager.
-            // For more information, see https://aka.ms/servicefabricactorsstateserialization
-
-            return this.StateManager.TryAddStateAsync("count", 0);
-        }
-
-        public async Task RecordEarningsJob(RecordEarningsJob message)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
