--- src/SFA.DAS.Payments.Monitoring.JobsService/PackageRoot/Config/Settings.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Settings xmlns:xsd="http://www.w3.org/2001/XMLSchema"
-          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-          xmlns="http://schemas.microsoft.com/2011/01/fabric">
-  <!-- The content will be generated during build -->
-</Settings>
