--- src/SFA.DAS.Payments.Monitoring.JobsService/PackageRoot/ServiceManifest.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ServiceManifest Name="SFA.DAS.Payments.Monitoring.JobsServicePkg"
-                 Version="1.0.0"
-                 xmlns="http://schemas.microsoft.com/2011/01/fabric"
-                 xmlns:xsd="http://www.w3.org/2001/XMLSchema"
-                 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <!-- The content will be generated during build -->
-</ServiceManifest>
