--- src/SFA.DAS.Payments.Monitoring.JobsService_old/Program.cs
@@ -1,28 +1,0 @@
-﻿using System;
-using System.Threading;
-using SFA.DAS.Payments.ServiceFabric.Core.Infrastructure.Ioc;
-
-namespace SFA.DAS.Payments.Monitoring.JobsService
-{
-    internal static class Program
-    {
-        /// <summary>
-        /// This is the entry point of the service host process.
-        /// </summary>
-        public static void Main()
-        {
-            try
-            {
-                using (ServiceFabricContainerFactory.CreateContainerForStatelessService<Monitoring.JobsService.JobsService>())
-                {
-                    Thread.Sleep(Timeout.Infinite);
-                }
-            }
-            catch (Exception e)
-            {
-                ServiceEventSource.Current.ServiceHostInitializationFailed(e.ToString());
-                throw;
-            }
-        }
-    }
-}
