--- src/SFA.DAS.Payments.Monitoring.JobsStatusService/PackageRoot/ServiceManifest.xml
@@ -1,22 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ServiceManifest xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema" Name="SFA.DAS.Payments.Monitoring.JobsStatusServicePkg" Version="1.0.0" xmlns="http://schemas.microsoft.com/2011/01/fabric">
-  <ServiceTypes>
-    <!-- This is the name of your ServiceType. 
-         This name must match the string used in the RegisterServiceAsync call in Program.cs. -->
-    <StatelessServiceType ServiceTypeName="SFA.DAS.Payments.Monitoring.JobsStatusServiceType" />
-  </ServiceTypes>
-  <CodePackage Name="Code" Version="1.0.0">
-    <EntryPoint>
-      <ExeHost>
-        <Program>SFA.DAS.Payments.Monitoring.JobsStatusService.exe</Program>
-      </ExeHost>
-    </EntryPoint>
-  </CodePackage>
-  <ConfigPackage Name="Config" Version="1.0.0" />
-  <Resources>
-    <Endpoints>
-      <Endpoint Name="ServiceEndpoint" />
-    </Endpoints>
-  </Resources>
-  <!-- The content will be generated during build -->
-</ServiceManifest>
