+++ src/SFA.DAS.Payments.Monitoring.ServiceFabric/ApplicationPackageRoot/ApplicationManifest.xml
@@ -1,6 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <ApplicationManifest xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" ApplicationTypeName="SFA.DAS.Payments.Monitoring.ServiceFabricType" ApplicationTypeVersion="1.0.0" xmlns="http://schemas.microsoft.com/2011/01/fabric">
   <Parameters>
+    <Parameter Name="SFA.DAS.Payments.Monitoring.Jobs.JobsProxyService_InstanceCount" DefaultValue="-1" />
     <Parameter Name="JobsProxyService_InstanceCount" DefaultValue="-1" />
     <Parameter Name="Monitoring_JobsService_InstanceCount" DefaultValue="-1" />
     <Parameter Name="Monitoring_JobsStatusService_InstanceCount" DefaultValue="1" />
@@ -17,36 +18,19 @@
     <Parameter Name="LogLevel" DefaultValue="Information" />
     <Parameter Name="ConsoleLogLevel" DefaultValue="" />
     <Parameter Name="DelayedRetryTimeInSeconds" DefaultValue="5" />
+    <Parameter Name="JobsServiceActorService_PartitionCount" DefaultValue="10" />
+    <Parameter Name="JobsServiceActorService_MinReplicaSetSize" DefaultValue="3" />
+    <Parameter Name="JobsServiceActorService_TargetReplicaSetSize" DefaultValue="3" />
   </Parameters>
   <!-- Import the ServiceManifest from the ServicePackage. The ServiceManifestName and ServiceManifestVersion 
        should match the Name and Version attributes of the ServiceManifest element defined in the 
        ServiceManifest.xml file. -->
   <ServiceManifestImport>
-    <ServiceManifestRef ServiceManifestName="SFA.DAS.Payments.Monitoring.JobsServicePkg" ServiceManifestVersion="1.0.0" />
+    <ServiceManifestRef ServiceManifestName="SFA.DAS.Payments.Monitoring.Jobs.JobsProxyServicePkg" ServiceManifestVersion="1.0.0" />
+    <ConfigOverrides />
   </ServiceManifestImport>
   <ServiceManifestImport>
-    <ServiceManifestRef ServiceManifestName="SFA.DAS.Payments.Monitoring.JobsProxyServicePkg" ServiceManifestVersion="1.0.0" />
-    <ConfigOverrides>
-      <ConfigOverride Name="Config">
-        <Settings>
-          <Section Name="ConnectionStringsSection">
-            <Parameter Name="StorageConnectionString" Value="[StorageConnectionString]" />
-            <Parameter Name="ServiceBusConnectionString" Value="[ServiceBusConnectionString]" />
-            <Parameter Name="LoggingConnectionString" Value="[LoggingConnectionString]" />
-            <Parameter Name="PaymentsConnectionString" Value="[PaymentsConnectionString]" />
-            <Parameter Name="PaymentsMonitoringConnectionString" Value="[PaymentsMonitoringConnectionString]" />
-          </Section>
-          <Section Name="Settings">
-            <Parameter Name="EndpointName" Value="[Monitoring_JobsService_EndpointName]" />
-            <Parameter Name="FailedMessagesQueue" Value="[Monitoring_JobsService_FailedMessagesQueue]" />
-            <Parameter Name="ApplicationInsightsInstrumentationKey" Value="[ApplicationInsightsInstrumentationKey]" />
-            <Parameter Name="DelayedRetryTimeInSeconds" Value="[DelayedRetryTimeInSeconds]" />
-            <Parameter Name="LogLevel" Value="[LogLevel]" />
-            <Parameter Name="ConsoleLogLevel" Value="[ConsoleLogLevel]" />
-          </Section>
-        </Settings>
-      </ConfigOverride>
-    </ConfigOverrides>
+    <ServiceManifestRef ServiceManifestName="SFA.DAS.Payments.Monitoring.Jobs.JobsServicePkg" ServiceManifestVersion="1.0.0" />
   </ServiceManifestImport>
   <DefaultServices>
     <!-- The section below creates instances of service types, when an instance of this 
@@ -54,10 +38,15 @@
          ServiceFabric PowerShell module.
          
          The attribute ServiceTypeName below must match the name defined in the imported ServiceManifest.xml file. -->
-    <Service Name="SFA.DAS.Payments.Monitoring.JobsProxyService" ServicePackageActivationMode="ExclusiveProcess">
-      <StatelessService ServiceTypeName="SFA.DAS.Payments.Monitoring.JobsProxyServiceType" InstanceCount="[JobsProxyService_InstanceCount]">
+    <Service Name="SFA.DAS.Payments.Monitoring.Jobs.JobsProxyService" ServicePackageActivationMode="ExclusiveProcess">
+      <StatelessService ServiceTypeName="SFA.DAS.Payments.Monitoring.Jobs.JobsProxyServiceType" InstanceCount="[SFA.DAS.Payments.Monitoring.Jobs.JobsProxyService_InstanceCount]">
         <SingletonPartition />
       </StatelessService>
+    </Service>
+    <Service Name="JobsServiceActorService" GeneratedIdRef="cb971f61-6bb9-4f53-a2d2-c61656f04754|Volatile">
+      <StatefulService ServiceTypeName="JobsServiceActorServiceType" TargetReplicaSetSize="[JobsServiceActorService_TargetReplicaSetSize]" MinReplicaSetSize="[JobsServiceActorService_MinReplicaSetSize]">
+        <UniformInt64Partition PartitionCount="[JobsServiceActorService_PartitionCount]" LowKey="-9223372036854775808" HighKey="9223372036854775807" />
+      </StatefulService>
     </Service>
   </DefaultServices>
 </ApplicationManifest>
